requirements.txt
streamlit
PyPDF2
langchain
langchain-core
langchain-community
langchain-anthropic
langchain-openai
langchain-google-genai==1.0.4
python-dotenv
----------------------------------------